+++ Acceso a datos/src/main/java/com/example/gym/service/ReservaServiceImpl.java
@@ -40,7 +40,7 @@ public class ReservaServiceImpl implements ReservaService {
     }
 
     @Override
-    public void deleteReserva(int id) {
+    public void deleteReserva(long id) {
         reservaRepository.findById(id)
                 .orElseThrow(() -> new ReservaNotFoundException(id));
         reservaRepository.deleteById(id);
